+++ gamestonk_terminal/helper_funcs.py
@@ -10,7 +10,6 @@ import re
 import sys
 import pandas as pd
 from pytz import timezone
-from prettytable import PrettyTable
 import iso8601
 import matplotlib
 import matplotlib.pyplot as plt
@@ -625,48 +624,6 @@ def plot_autoscale():
     return x, y
 
 
-def print_and_record_reddit_post(submissions_dict, submission):
-    # Refactor data
-    s_datetime = datetime.utcfromtimestamp(submission.created_utc).strftime(
-        "%d/%m/%Y %H:%M:%S"
-    )
-    s_link = f"https://old.reddit.com{submission.permalink}"
-    s_all_awards = ""
-    for award in submission.all_awardings:
-        s_all_awards += f"{award['count']} {award['name']}\n"
-    s_all_awards = s_all_awards[:-2]
-    # Create dictionary with data to construct dataframe allows to save data
-    submissions_dict[submission.id] = {
-        "created_utc": s_datetime,
-        "subreddit": submission.subreddit,
-        "link_flair_text": submission.link_flair_text,
-        "title": submission.title,
-        "score": submission.score,
-        "link": s_link,
-        "num_comments": submission.num_comments,
-        "upvote_ratio": submission.upvote_ratio,
-        "awards": s_all_awards,
-    }
-    # Print post data collected so far
-    print(f"{s_datetime} - {submission.title}")
-    print(f"{s_link}")
-    t_post = PrettyTable(
-        ["Subreddit", "Flair", "Score", "# Comments", "Upvote %", "Awards"]
-    )
-    t_post.add_row(
-        [
-            submission.subreddit,
-            submission.link_flair_text,
-            submission.score,
-            submission.num_comments,
-            f"{round(100 * submission.upvote_ratio)}%",
-            s_all_awards,
-        ]
-    )
-    print(t_post)
-    print("\n")
-
-
 def get_last_time_market_was_open(dt):
     # Check if it is a weekend
     if dt.date().weekday() > 4:
@@ -679,20 +636,6 @@ def get_last_time_market_was_open(dt):
     dt = dt.replace(hour=21, minute=0, second=0)
 
     return dt
-
-
-def find_tickers(submission):
-    ls_text = [submission.selftext, submission.title]
-    submission.comments.replace_more(limit=0)
-    for comment in submission.comments.list():
-        ls_text.append(comment.body)
-
-    l_tickers_found = []
-    for s_text in ls_text:
-        for s_ticker in set(re.findall(r"([A-Z]{3,5} )", s_text)):
-            l_tickers_found.append(s_ticker.strip())
-
-    return l_tickers_found
 
 
 def export_data(
